Reduce test duplication and consolidate test utility files
