+++ workspaces/x2a/plugins/x2a-backend/src/services/X2ADatabaseService/jobsUpdateDelete.test.ts
@@ -26,10 +26,10 @@ import {
   createDatabase,
   createService,
   defaultProjectRepoFields,
+  nonExistentId,
   supportedDatabaseIds,
   tearDownDatabases,
-} from './__testUtils__/testHelpers';
-import { nonExistentId } from '../../utils/tests';
+} from '../../__testUtils__';
 
 describe('X2ADatabaseService – jobs (update & delete)', () => {
   afterEach(async () => {
