+++ workspaces/x2a/plugins/x2a-backend/src/services/X2ADatabaseService/projectsGet.test.ts
@@ -21,10 +21,10 @@ import {
   createDatabase,
   createService,
   defaultProjectRepoFields,
+  nonExistentId,
   supportedDatabaseIds,
   tearDownDatabases,
-} from './__testUtils__/testHelpers';
-import { nonExistentId } from '../../utils/tests';
+} from '../../__testUtils__';
 
 describe('X2ADatabaseService – projects (get & delete)', () => {
   afterEach(async () => {
